feat(registry): registry manages send channels
To ensure that there are no dropped metrics in the system when we
update/delete objects, the registry will manage the channel lifecycle
and associate both the collector and discovery services to them.  The
registry has been broken up into 2 parts.  The manager which is exposed
to the controller, giving the entrypoints/abstractions to add and delete
and the registry operations.  This keeps us from any circular
dependencies as we expose the registry to both the discovery and
collextion services.  The registry will create a send channel on the
addition of any new collector and remove that channel when the collector
has been removed by the controller.  It remains intact during updates
where the original underlying object is being rotated out.