feat(collector): configurable buffer size
Adding the buffer size into the collector.  Even after the channels were
moved to the registry, it makes sense (at least to me) to keep the
configuration in the collector.  Making a seperate global config would
be yet another thing to manage - not that it's a bad thing, but I want
to keep things as simple as possible.  But, even in the registry, the
channels are still tied to the collectors, so it's still related.