feat(filters): add filters
Added filters which will run before the metrics are sent to the output.
Currently there are clip and exclude.  In addition, a new type "Mutator"
was set up for a special type that will mutate the metric.  The clamp
filter will belong to this group, plus any others that will modify the
metric in time