chore(all): add license headers
Mostly adds license headers, but realized that the client generation was
broken and not occuring in the codegen target.  Even though the
code-generator bin was creating the deepcopy methods, the client stuff
was missed.  We don't actually use the client since this crd isn't
necessarily used to provide an interface into other programatically
managed environments, however we might actually use it once we start
testing, so it's been fixed.